backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PDF Generation
reportlab>=4.0.7
Pillow>=10.3.0
pypdf2>=3.0.1

# HTTP Client
httpx>=0.25.1
aiofiles>=23.2.1

# Network Storage
webdavclient3>=3.14.6
smbprotocol>=1.12.0
aioftp>=0.21.4

# Configuration
pydantic>=2.5.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
